Question 1.
Attached a click event listener to the button (#submit). When the button is clicked, the listener constructs a URL to the text file data.txt, creates an instance of XMLHttpRequest, configures it to retrieve data.txt via an asynchronous GET request, and in the callback function, sets the text content of the paragraph (#result) to the response text.